identify ERC20, ERC721, ERC1155 transfers, and save them in different tables

+++ src/main/java/com/xkb/web3j/entity/CustomErc20Transfer.java
@@ -3,6 +3,8 @@ package com.xkb.web3j.entity;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
+
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.io.Serializable;
 import io.swagger.annotations.ApiModel;
@@ -47,13 +49,13 @@ public class CustomErc20Transfer implements Serializable {
     private String txHash;
 
     @ApiModelProperty(value = "交易序号（在区块中）")
-    private Long txIndex;
+    private BigInteger txIndex;
 
     @ApiModelProperty(value = "区块哈希值")
     private String blockHash;
 
     @ApiModelProperty(value = "区块编号")
-    private Long blockNumber;
+    private BigInteger blockNumber;
 
     @ApiModelProperty(value = "创建时间")
     private LocalDateTime createTime;
